Add nashorn as external dependency
This adds dependency management through ivy (still needs to move prior two deps to be managed with ivy), it also adds nashorn and all its depending stuff to classpath.

Adds a fix from scriptcraft to the Java Files - didn't check if that really is necessary, but at least it didn't break anything.